Move all business logic and file write-read to service.js
